refactor: #546 DiscussionRoom 패키지 변경 및 Title, Opinion 이름 변경
