+++ android/TodokTodok/domain/src/main/java/com/team/domain/model/discussionroom/Opinion.kt
@@ -1,4 +1,4 @@
-package com.team.domain.model.member
+package com.team.domain.model.discussionroom
 
 @JvmInline
 value class Opinion(
